create a new bellman_ford module
